Release blind at end position
In non-end positions we end our transmission with a "stop" signal. This
means that other controls (such as manual switches) can now interact
with the blinds as well.

In end positions, we did not send a "stop" signal to ensure that the
blinds can recover from invalid state. However, that means that we may
now be in a situation where manual controls no longer work.

So let's send a stop signal for end positions too, but simply do that
after a full blind closure cycle has elapsed. That way we should never
interfere with the recovery part of end positions, but still free the
blinds for manual control.

Signed-off-by: Alexander Graf <agraf@csgraf.de>